PyTorch/examples/huggingface/diffusers/requirements.txt
optimum-habana==1.11.1
